+++ daengle-groomer-api/src/main/java/ddog/groomer/application/EstimateService.java
@@ -2,8 +2,9 @@ package ddog.groomer.application;
 
 import ddog.domain.estimate.GroomingEstimate;
 import ddog.domain.estimate.dto.request.GroomerGroomingEstimateReq;
-import ddog.domain.estimate.dto.response.GroomingEstimateInfo;
-import ddog.domain.estimate.dto.response.UserGroomingEstimateDetail;
+import ddog.groomer.application.mapper.GroomingEstimateMapper;
+import ddog.groomer.presentation.estimate.dto.GroomingEstimateInfo;
+import ddog.groomer.presentation.estimate.dto.GroomingEstimateDetail;
 import ddog.domain.groomer.Groomer;
 import ddog.persistence.port.GroomerPersist;
 import ddog.persistence.port.GroomingEstimatePersist;
@@ -25,13 +26,13 @@ public class EstimateService {
         Groomer groomer = groomerPersist.getGroomerByAccountId(accountId);
         List<GroomingEstimate> generalEstimates = groomingEstimatePersist.findGeneralGroomingEstimates(groomer.getAddress());
         List<GroomingEstimate> designationEstimates = groomingEstimatePersist.findDesignationGroomingEstimates(groomer.getGroomerId());
-        return GroomingEstimate.toGroomingEstimateInfo(generalEstimates, designationEstimates);
+        return GroomingEstimateMapper.toGroomingEstimateInfo(generalEstimates, designationEstimates);
     }
 
     @Transactional(readOnly = true)
-    public UserGroomingEstimateDetail getGroomingEstimateDetailInfo(Long groomingEstimateId) {
+    public GroomingEstimateDetail getGroomingEstimateDetailInfo(Long groomingEstimateId) {
         GroomingEstimate groomingEstimate = groomingEstimatePersist.getByGroomingEstimateId(groomingEstimateId);
-        return groomingEstimate.toUserGroomingEstimateDetail();
+        return GroomingEstimateMapper.toGroomingEstimateDetail(groomingEstimate);
     }
 
     @Transactional
