+++ daengle-user-api/src/main/java/ddog/user/presentation/account/dto/UserInfo.java
@@ -1,4 +1,4 @@
-package ddog.domain.estimate.dto.response;
+package ddog.user.presentation.account.dto;
 
 import ddog.domain.pet.Weight;
 import lombok.Builder;
